move stuff to shared folder
