+++ frontend/src/components/Shared/Map/AddCrossingMap.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import ReactMapboxGl, { Layer, Feature } from "react-mapbox-gl";
-import mapboxstyle from 'components/Map/mapboxstyle.json';
+import mapboxstyle from 'components/Shared/Map/mapboxstyle.json';
 
 const Map = ReactMapboxGl({ accessToken: null });
