+++ frontend/src/components/Shared/Map/CrossingMap.scss
