+++ frontend/src/components/Shared/Map/CrossingStaticMap.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import ReactMapboxGl, { Marker } from "react-mapbox-gl";
-import mapboxstyle from 'components/Map/mapboxstyle.json';
+import mapboxstyle from 'components/Shared/Map/mapboxstyle.json';
 import { statusIcons } from 'constants/StatusConstants';
 
 const Map = ReactMapboxGl({ accessToken: null, interactive: false });
