+++ frontend/src/components/Shared/Map/style/scripts/build.js
