+++ frontend/src/components/Shared/Map/style/scripts/check-fonts.js
