+++ frontend/src/components/Shared/Map/style/scripts/lang-fallback.js
