+++ frontend/src/components/Shared/Map/style/scripts/validate.js
