+++ frontend/src/components/Shared/Map/style/style.json
